Solved the bugs explained in bug/bug.md
All the print statements used to solve the bugs were removed.

#######  search.py  #######
Updates in Query

To fix the first bug, which assigned the between type to all date searches.

When checking the date flags to see if we want between or equals, instead of checking whether the dictionary contains the start_date and end_date keys, we check whether the keys are None.

#######  database.py  #######
Updates in Database

The function load_data was not converting the string date to a datetime object. Therefore, comparing dates in search.py always returned False. Now it is converted to a datetime object.

#######  search.py  ######
Updates in NEOSearcher.

The third bug was the search was returning all occurrences of a given date search, instead of the number of wanted occurrences.

To solve this, the while loop in line 45 was removed. Instead, an if statement was added that checks that the number of found items is lower than required. When it is as required, it breaks the for loop.