Solved requirment 1
Had to do a few changes to pass the requirments:
    - changed the output of get_objects to a list rather than a dict.
    - when checking which type of date it is:
      - first check that the start and end date are in the dict keys
      - then check that their value is not None

Strategy:
    - Construct the query (with empty filters):
      - determine the type of date and convert it to datetime
      - determine the return object

    - Pass the query to the get_objects function:
      - initialise variable found (for found items) to 0
      - for all the objects in the database:
        - if any of the dates in orbit_dates passes the date requirments given in query
        - get the appropiate return_object and add it to an output list.
        - increase found by one
        - if we have found enough items: stop and return the list

The strategy to construct the Query is straightforward. The namedtuples were given, so I just had to make sure to add the appropiate information to them.

In order to check what kind of date is needed there are checks:
   - whether the start_date and end_date flags are keys in kwargs
   - whether their values are None or not
   this is because, depending on where the function is called from, kwargs may have the non-given flags as (key, None) pairs, or they might not exist at all.

The search strategy required more thought. The reason to have a counter for the found items is that, in most cases, the program will not have to search through the entire database.

Another key point was to make sure to only retrieve the appropiate return_object when there was a match in the dates, to avoid redundant computation.

Data Structures:

Query:

Query.Selectors: namedtuple where:
    - date_search: Query.DateSearch namedtuple, where:
      - type is either equals or between
      - values is a list with the date values in ascending order
    - number: int
    - filters: empty list for now
    - return_object: either NearEarthObject or OrbitPath

Using namedtuples to structure the Query.Selectors seems reasonable as it is a good way to keep the selectors organised and easily indexable.

NEOSearcher:

    - the output is simply a list:
      - it needs to be mutable as items have to be appended
      - it should also be unique, but there is only one iteration through the database,
        which already guarantees uniqueness for each name. Therefore, there was not
        much difference between using a list or a set.

Database:
    The database is simply a dictionary, where eaxh key represents the name of a NEO and each value is a NEO.
    Each NEO has a class attribute orbits: a list with all its OrbitPaths.