Solved requirment 2 and 3
There was no specific unit test for requirment 2. Therefore, requirments 2 and 3 are joined in this commit.

######## Query #######
Updated Query to have an extra class attribute:
    - filter: list of filters as str

Build query also returns these filter values in Query.Selectors.filters

######## Filters ########

Started implementing the Filters class. Did the basic functionality.

Class attributes:
    - field (str): field to perform the filter on
    - object (str): object to perform the filter on
    - operation (str): operation needed for the filter
    - value (str): value wanted

Functions:
    - create_filter_options(filter_options)
      Takes a list of strings, each represent a filter. Returns a list (see data structures)
      of filters obtained from calling _parse_filter with each string.

    - _parse_filter(filter_str)
      The diameter can have operators <, <=, > and >=. The database does not provide an
      exact diameter, but upper and lower boundaries for it. Therefore, in order to
      guarantee correct results, when requesting diameters larger-than the upper
      boundary is used; similarly, for diameters lower-than the lower boundary is used.
      For equal the upper one is used.
      In order to make the distinction, if the operations symbols are >, >=, or =,
      diam_max is used as the field.

    - apply(results)
      Instance method that takes a list of results and applies its filter.
      Returns a list made of the filter function that filters results using the
      function _filter_p.

    - _filter_p(neo)
      Takes a neo and checkes whether it passes the filter. Returns a boolean.
      It is divided in two sections
         - For filters on NearEarthObjects:
           returns the output of the filter's operation, the NEO's attribute (field) and
           the filter's value.
         - For filters on OrbitPath:
           returns True if any of the OrbitPath object in neo pass the filter.

For performing filters in the diameters, the diameter_km_min was chosen to make it consistent with the unit tests.

######## NEOSearcher #######

There were several changes from the last commit:

The found counter disappears in order to merge the search with the logic of the filters

The new search is as follows:
    - get a list of filters given by Filter.create_filter_options, passing as argument the
      list of strings in query.filter
    - get the results that match the date requirments and store them in a list results
    - for all filters:
      - update results to the result of filter.apply(results)
    - if the return object is OrbitPath:
      - get the appropiate OrbitPath from the NEO.

######## Data Structures #######

#### Filter

There are two class variables:
      - Options: dictionary where
        - key: field to filter on
        - value: (NearEarthObject/OrbitPath, attribute

      - Operators: dict where
        - key: operation symbol
        - value: correconding function from library operators

The values in Options are structured as such because the object that the attribute corresponds to is needed to perform the filter.

Another data structure worth mentioning and justifying is the output of create_filter_options. The function was previously defined as returning a defaultdict, where the keys were NearEarthObject/OrbitPath and the values lists of filters. However, each filter already contains which object it applies to as a class attribute, and the implementation of Filter._filter_p, the function where filters are tested, is implemented such that only this information is needed. Finally, this data_structure made the application of filters in NEOSearcher.get_objects much simpler.

#### NEOSearcher

There was not much variation in data structures. The results are kept in a list, as it has to be mutated and it  will not have repeated elements due it having only one for loop on the database.

#### Unit Tests

The names of the class attributes were changed in the unit tests. No functionallity or testing is new. Simply, the attributes for the diameter and potentially hazardous were named different in my implementation of models.py and in the Unit Tests.