Solved requirment 4
Solving this requirment only needed very minor changes from the previous implementation.
The only thing that had to be done was add to Filter.Options a new entry
    - 'distance': ('OrbitPath', 'miss')

#### Unit Tests

The names of the class attributes were changed in the unit tests. No functionallity or testing is new. Simply, the attributes for the miss distance and names were named different in my implementation of models.py and in the Unit Tests.